feat(home): Added Sparkles to the main section.
Sparkles were added to the main section, and the Section component was
changed to use an actual section.

fix(home): Changed imports to use View component. Updated canvas z-index.

View component wasn't importing, had to set the actual module in the
dynamicimport for the Noiseshader. Then had a actual reason to use
lenis (It wasn't enabled as it turns out on touchscreen), as iOS safari
still suffers from the jitter bugs.

With the move towards View components, it was necessary to set the z-index
on the canvas, since it wraps the majority of the dom.