feat: fumadocs integration on atom.xml route + pages + post-previews
Atom.xml endpoint was refactored to integrate with the fumadocs-mdx
sources. A redundant source for the blog-posts was created, in order to
have separate mdx-compilation plugins (rehype-highlight on blog
collection, no rehype-code plugin for blogfeed collection).

Additionally, the html for the atom.xml route is rendered on the server.
This is achieved via a try-catch async function, which seems to allow
renderToStaticMarkup to fail quietly. This essentially skips over imported
client components in the mdx (so long as they are wrapped in a suspense
boundary in a server component) which is necessary for this use case, as
it allows for the rest of the page to be rendered anyways.